Fix some clang-tidy warnings
Incorporated some fixes by Daniel Chabrowski (#467)

+++ src/messages/initialize.cc
@@ -370,7 +370,7 @@ void do_initialize(MessageHandler *m, InitializeParam &param,
   // may take a long time. Indexer threads will emit status/progress
   // reports.
   if (g_config->index.threads == 0)
-    g_config->index.threads = std::thread::hardware_concurrency();
+    g_config->index.threads = (int)std::thread::hardware_concurrency();
 
   LOG_S(INFO) << "start " << g_config->index.threads << " indexers";
   for (int i = 0; i < g_config->index.threads; i++)
